fix: fallback to mha without avx512f support
